refact: move use cases
