+++ src/input/view-controllers/IndexViewController.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
 import { Transaction } from "../../application/domain/Transaction";
-import { FindAllFromUserUseCase } from "../../application/useCases/FindAllFromUserUseCase";
+import { LoadAllTransactionsFromUser } from "../../application/useCases/transactions/LoadAllTransactionsFromUser";
+
 import IRepository from "../../output/repositories/IRepository";
 
 export class IndexViewController {
@@ -8,7 +9,7 @@ export class IndexViewController {
     constructor(private readonly repository: IRepository<Transaction>) { };
 
     async handle(request: Request, response: Response) {
-        const transactions = await new FindAllFromUserUseCase(this.repository).execute("");
+        const transactions = await new LoadAllTransactionsFromUser(this.repository).execute("");
         return response.render("index", { user: "No User", transactions });
     }
 }
